crawler/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
psycopg[binary]==3.2.3
anthropic==0.25.0
openai==1.30.0
python-dotenv==1.0.1
click==8.1.7
praw==7.7.1
pytest==8.2.0
pytest-asyncio==0.23.6
